backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
openai==1.3.0
google-generativeai==0.3.0
pillow==10.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
requests==2.31.0
